Client-Side  + Renamed Main to SeriousClient  + Started back-end skeleton for P2P networking  + Changed client-side hierarchy:    - Client extends User extends ContactData    - Contact extends ContactData  + Added disconnect handler for connected contacts  + Modified signal system so that P2P and server signals are split up
Server-Side
 + Renamed Main to SeriousServer
 + Added ability to delete users from the database
 + Added confirm dialogs to database table operations (ie. create / delete / reset)

git-svn-id: file:///e/SVN/serious_messenger@43 b1debafe-a605-4e1b-80f9-83a58a4e6cd1

+++ Serious Client/src/client/ClientState.java
@@ -7,7 +7,7 @@ public class ClientState {
 	final public static byte AwaitingAuthentication = 2;
 	final public static byte Online = 3;
 	
-	public static boolean isValid(int state) {
+	public static boolean isValid(byte state) {
 		return state >= 0 && state <= 3;
 	}
